backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic>=2.0
